move jobEventRepository functions to JobEventQuery #5748
