+++ src/Repository/JobEventRepository.php
@@ -2,16 +2,8 @@
 
 namespace App\Repository;
 
-use App\Entity\Enum\InterfacageType;
 use App\Entity\JobEvent;
-use App\Entity\Partner;
-use App\Entity\Signalement;
-use App\Repository\Behaviour\EntityCleanerRepositoryInterface;
-use App\Service\ListFilters\SearchInterconnexion;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\ORM\NonUniqueResultException;
-use Doctrine\ORM\NoResultException;
-use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
 
 /**
@@ -22,201 +14,10 @@ use Doctrine\Persistence\ManagerRegistry;
  * @method JobEvent[]    findAll()
  * @method JobEvent[]    findBy(array<string, mixed> $criteria, array<string, mixed>|null $orderBy = null, $limit = null, $offset = null)
  */
-class JobEventRepository extends ServiceEntityRepository implements EntityCleanerRepositoryInterface
+class JobEventRepository extends ServiceEntityRepository
 {
     public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, JobEvent::class);
-    }
-
-    /**
-     * @throws \DateMalformedStringException
-     */
-    private function createJobEventByTerritoryQueryBuilder(
-        int $dayPeriod,
-        SearchInterconnexion $searchInterconnexion,
-        bool $count = false,
-    ): QueryBuilder {
-        $qb = $this
-            ->createQueryBuilder('j')
-            ->where('j.createdAt >= :date_limit')
-            ->setParameter('date_limit', new \DateTimeImmutable('-'.$dayPeriod.' days'));
-
-        $needsPartnerJoin =
-            null !== $searchInterconnexion->getTerritory()
-            || null !== $searchInterconnexion->getPartner()
-            || !$count;
-
-        $needsSignalementJoin =
-            null !== $searchInterconnexion->getReference()
-            || !$count;
-
-        if ($needsPartnerJoin) {
-            $joinType = ($searchInterconnexion->getTerritory() || $searchInterconnexion->getPartner())
-                ? 'innerJoin'
-                : 'leftJoin';
-
-            $qb->{$joinType}(Partner::class, 'p', 'WITH', 'p.id = j.partnerId');
-        }
-
-        if ($needsSignalementJoin) {
-            $joinType = $searchInterconnexion->getReference()
-                ? 'innerJoin'
-                : 'leftJoin';
-
-            $qb->{$joinType}(Signalement::class, 's', 'WITH', 's.id = j.signalementId');
-        }
-
-        if ($searchInterconnexion->getTerritory()) {
-            $qb->andWhere('p.territory = :territory')->setParameter('territory', $searchInterconnexion->getTerritory());
-        }
-        if ($searchInterconnexion->getPartner()) {
-            $qb->andWhere('p.id = :partnerId')->setParameter('partnerId', $searchInterconnexion->getPartner()->getId());
-        }
-        if ($searchInterconnexion->getStatus()) {
-            if ('failed' === $searchInterconnexion->getStatus()) {
-                $qb->andWhere('j.status = :status')->setParameter('status', JobEvent::STATUS_FAILED);
-            } elseif ('warning' === $searchInterconnexion->getStatus()) {
-                $qb
-                    ->andWhere('j.status = :status')
-                    ->andWhere('(
-                        j.response IS NULL
-                        OR j.response = \'{}\'
-                        OR j.response = \'[]\'
-                        OR LENGTH(TRIM(j.response)) = 0
-                    )')
-                    ->setParameter('status', JobEvent::STATUS_SUCCESS);
-            } else {
-                $qb->andWhere('j.status = :status')
-                    ->andWhere('j.response IS NOT NULL')
-                    ->andWhere('LENGTH(TRIM(j.response)) > 0')
-                    ->andWhere('j.response != \'[]\'')
-                    ->andWhere('j.response != \'{}\'')
-                    ->setParameter('status', $searchInterconnexion->getStatus())
-                ;
-            }
-        }
-        if ($searchInterconnexion->getAction()) {
-            $qb->andWhere('j.action = :action')->setParameter('action', $searchInterconnexion->getAction());
-        }
-        if ($searchInterconnexion->getReference()) {
-            $qb->andWhere('s.reference = :reference')->setParameter('reference', $searchInterconnexion->getReference());
-        }
-
-        if ($searchInterconnexion->getService()) {
-            $qb->andWhere('j.service = :service')->setParameter('service', $searchInterconnexion->getService());
-        }
-
-        return $qb;
-    }
-
-    /**
-     * @return array<int, array<string, mixed>>
-     *
-     * @throws \DateMalformedStringException
-     */
-    public function findLastJobEventByTerritory(
-        int $dayPeriod,
-        SearchInterconnexion $searchInterconnexion,
-        int $limit,
-        int $offset,
-    ): array {
-        /** @var QueryBuilder $qb */
-        $qb = $this->createJobEventByTerritoryQueryBuilder($dayPeriod, $searchInterconnexion);
-        $qb->select(
-            'j.createdAt',
-            'p.id',
-            'p.nom',
-            's.reference',
-            's.uuid',
-            'j.status',
-            'j.service',
-            'j.action',
-            'j.codeStatus',
-            'j.message',
-            'j.response',
-            'j.attachmentsCount',
-            'j.attachmentsSize'
-        );
-
-        if (!empty($searchInterconnexion->getOrderType())) {
-            [$orderField, $orderDirection] = explode('-', $searchInterconnexion->getOrderType());
-            $qb->orderBy($orderField, $orderDirection);
-        } else {
-            $qb->orderBy('j.createdAt', 'DESC');
-        }
-
-        $qb->setMaxResults($limit)->setFirstResult($offset);
-
-        return $qb->getQuery()->getArrayResult();
-    }
-
-    /**
-     * @throws NonUniqueResultException
-     * @throws NoResultException
-     * @throws \DateMalformedStringException
-     */
-    public function countLastJobEventByTerritory(
-        int $dayPeriod,
-        SearchInterconnexion $searchInterconnexion,
-    ): int {
-        $qb = $this->createJobEventByTerritoryQueryBuilder($dayPeriod, $searchInterconnexion, true);
-        $qb->select('COUNT(j.id)');
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
-    /**
-     * @return array<string, mixed>|null
-     *
-     * @throws NonUniqueResultException
-     */
-    public function findLastEsaboraJobByPartner(
-        Partner $partner,
-    ): ?array {
-        return $this->createQueryBuilder('j')
-            ->select('MAX(j.createdAt) AS last_event')
-            ->innerJoin(Partner::class, 'p', 'WITH', 'p.id = j.partnerId')
-            ->where('p.id = :partner')->setParameter('partner', $partner->getId())
-            ->andWhere('j.service = :service')
-            ->setParameter('service', InterfacageType::ESABORA->value)
-            ->getQuery()
-            ->getOneOrNullResult();
-    }
-
-    /**
-     * @throws \Exception
-     */
-    public function cleanOlderThan(string $periodFailed = JobEvent::EXPIRATION_PERIOD_FAILED): int
-    {
-        $periodDefault = JobEvent::EXPIRATION_PERIOD_DEFAULT;
-        $queryBuilder = $this->createQueryBuilder('j');
-        $queryBuilder->delete()
-            ->andWhere('(DATE(j.createdAt) <= :created_at AND j.status = :status_failed) OR (DATE(j.createdAt) <= :created_at_default AND j.status != :status_failed)')
-            ->setParameter('created_at', (new \DateTimeImmutable($periodFailed))->format('Y-m-d'))
-            ->setParameter('created_at_default', (new \DateTimeImmutable($periodDefault))->format('Y-m-d'))
-            ->setParameter('status_failed', JobEvent::STATUS_FAILED);
-
-        return $queryBuilder->getQuery()->execute();
-    }
-
-    /**
-     * @return array<string, int>
-     */
-    public function getReportEsaboraAction(string ...$actions): array
-    {
-        $qb = $this->createQueryBuilder('j');
-        $qb->select([
-            'SUM(CASE WHEN j.status = :status_success THEN 1 ELSE 0 END) AS success_count',
-            'SUM(CASE WHEN j.status = :status_failed THEN 1 ELSE 0 END) AS failed_count',
-        ])
-            ->andWhere('j.action IN (:actions)')
-            ->andWhere('DATE(j.createdAt) = :today')
-            ->setParameter('actions', $actions)
-            ->setParameter('today', (new \DateTimeImmutable())->format('Y-m-d'))
-            ->setParameter('status_success', JobEvent::STATUS_SUCCESS)
-            ->setParameter('status_failed', JobEvent::STATUS_FAILED);
-
-        return $qb->getQuery()->getSingleResult();
     }
 }
